+++ ReactViewResources/Loader/Task.ts
@@ -1,0 +1,17 @@
+﻿export class Task<ResultType> {
+
+    private taskPromise: Promise<ResultType>;
+    private resolve: (result: ResultType) => void;
+
+    constructor() {
+        this.taskPromise = new Promise<ResultType>((resolve) => this.resolve = resolve);
+    }
+
+    public setResult(result?: ResultType) {
+        this.resolve(result as ResultType);
+    }
+
+    public get promise() {
+        return this.taskPromise;
+    }
+}
